extension/CSV_TCP_SENDER/pyproject.toml
[project]
name = "csv-tcp-sender"
version = "0.1.0"
description = "CSV data to TCP gateway sender"
readme = "README.md"
requires-python = ">=3.10.18"
dependencies = [
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
test = [
    "pytest>=8.4.2",
]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
